# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeoTrack224 Intelligence API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeviceType = (typeof DeviceType)[keyof typeof DeviceType];

export const DeviceType = {
  vehicle: "vehicle",
  asset: "asset",
  person: "person",
  drone: "drone",
} as const;

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  moving: "moving",
  stopped: "stopped",
  offline: "offline",
} as const;

export interface Device {
  id: number;
  name: string;
  type: DeviceType;
  status: DeviceStatus;
  imei: string;
  speedLimit?: number;
  groupName?: string;
  groupId?: number;
  userId?: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateDeviceBodyType =
  (typeof CreateDeviceBodyType)[keyof typeof CreateDeviceBodyType];

export const CreateDeviceBodyType = {
  vehicle: "vehicle",
  asset: "asset",
  person: "person",
  drone: "drone",
} as const;

export interface CreateDeviceBody {
  name: string;
  type: CreateDeviceBodyType;
  imei: string;
  speedLimit?: number;
  groupName?: string;
  groupId?: number;
  userId?: number;
}

export type UpdateDeviceBodyType =
  (typeof UpdateDeviceBodyType)[keyof typeof UpdateDeviceBodyType];

export const UpdateDeviceBodyType = {
  vehicle: "vehicle",
  asset: "asset",
  person: "person",
  drone: "drone",
} as const;

export interface UpdateDeviceBody {
  name?: string;
  type?: UpdateDeviceBodyType;
  speedLimit?: number;
  groupName?: string;
  groupId?: number;
  userId?: number;
}

export interface AssignDeviceBody {
  groupId?: number | null;
  userId?: number | null;
}

export interface Position {
  id: number;
  deviceId: number;
  latitude: number;
  longitude: number;
  speed: number;
  heading: number;
  altitude?: number;
  accuracy?: number;
  timestamp: string;
}

export type LivePositionDeviceType =
  (typeof LivePositionDeviceType)[keyof typeof LivePositionDeviceType];

export const LivePositionDeviceType = {
  vehicle: "vehicle",
  asset: "asset",
  person: "person",
  drone: "drone",
} as const;

export type LivePositionStatus =
  (typeof LivePositionStatus)[keyof typeof LivePositionStatus];

export const LivePositionStatus = {
  moving: "moving",
  stopped: "stopped",
  offline: "offline",
} as const;

export interface LivePosition {
  deviceId: number;
  deviceName: string;
  deviceType: LivePositionDeviceType;
  status: LivePositionStatus;
  latitude: number;
  longitude: number;
  speed: number;
  heading: number;
  timestamp: string;
  groupName?: string;
}

export interface IngestPositionBody {
  deviceId: number;
  latitude: number;
  longitude: number;
  speed: number;
  heading: number;
  altitude?: number;
  accuracy?: number;
  timestamp?: string;
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  overspeed: "overspeed",
  offline: "offline",
  geofence: "geofence",
} as const;

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  deviceId: number;
  deviceName?: string;
  type: AlertType;
  message: string;
  severity: AlertSeverity;
  acknowledged: boolean;
  createdAt: string;
  acknowledgedAt?: string;
}

export interface DashboardSummary {
  totalDevices: number;
  movingCount: number;
  stoppedCount: number;
  offlineCount: number;
  activeAlerts: number;
  vehicleCount: number;
  assetCount: number;
  personCount: number;
  droneCount: number;
  totalPositionsToday: number;
}

export type ActivityEventType =
  (typeof ActivityEventType)[keyof typeof ActivityEventType];

export const ActivityEventType = {
  position_update: "position_update",
  status_change: "status_change",
  alert_triggered: "alert_triggered",
  alert_acknowledged: "alert_acknowledged",
} as const;

export interface ActivityEvent {
  id: number;
  deviceId: number;
  deviceName: string;
  type: ActivityEventType;
  description: string;
  timestamp: string;
}

export interface Group {
  id: number;
  name: string;
  description?: string;
  color: string;
  deviceCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGroupBody {
  name: string;
  description?: string;
  color: string;
}

export interface UpdateGroupBody {
  name?: string;
  description?: string;
  color?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  operator: "operator",
  viewer: "viewer",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  groupId?: number;
  groupName?: string;
  deviceCount?: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  operator: "operator",
  viewer: "viewer",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  role?: CreateUserBodyRole;
  groupId?: number;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  operator: "operator",
  viewer: "viewer",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  groupId?: number;
}

export type ListDevicesParams = {
  type?: ListDevicesType;
  status?: ListDevicesStatus;
};

export type ListDevicesType =
  (typeof ListDevicesType)[keyof typeof ListDevicesType];

export const ListDevicesType = {
  vehicle: "vehicle",
  asset: "asset",
  person: "person",
  drone: "drone",
} as const;

export type ListDevicesStatus =
  (typeof ListDevicesStatus)[keyof typeof ListDevicesStatus];

export const ListDevicesStatus = {
  moving: "moving",
  stopped: "stopped",
  offline: "offline",
} as const;

export type GetDeviceHistoryParams = {
  limit?: number;
  from?: string;
  to?: string;
};

export type ListAlertsParams = {
  acknowledged?: boolean;
  type?: ListAlertsType;
  limit?: number;
};

export type ListAlertsType =
  (typeof ListAlertsType)[keyof typeof ListAlertsType];

export const ListAlertsType = {
  overspeed: "overspeed",
  offline: "offline",
  geofence: "geofence",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};