# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GeoTrack224 Intelligence API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: devices
    description: GPS device management
  - name: positions
    description: GPS position tracking
  - name: alerts
    description: Alert management
  - name: stats
    description: Dashboard statistics
  - name: groups
    description: Group management
  - name: users
    description: User management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all tracked devices
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [vehicle, asset, person, drone]
        - name: status
          in: query
          schema:
            type: string
            enum: [moving, stopped, offline]
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new GPS device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceBody"
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Update a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceBody"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /devices/{id}/assign:
    put:
      operationId: assignDevice
      tags: [devices]
      summary: Assign a device to a group and/or user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDeviceBody"
      responses:
        "200":
          description: Device assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found

  /devices/{id}/positions:
    get:
      operationId: getDeviceHistory
      tags: [positions]
      summary: Get position history for a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Position history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"

  /positions:
    post:
      operationId: ingestPosition
      tags: [positions]
      summary: Ingest a GPS position update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestPositionBody"
      responses:
        "201":
          description: Position recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"

  /positions/live:
    get:
      operationId: getLivePositions
      tags: [positions]
      summary: Get latest position for all devices
      responses:
        "200":
          description: Latest positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LivePosition"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: acknowledged
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
            enum: [overspeed, offline, geofence]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  /stats/summary:
    get:
      operationId: getDashboardSummary
      tags: [stats]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /stats/activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"

  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"

  /groups/{id}:
    put:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupBody"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Not found
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Device:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [vehicle, asset, person, drone]
        status:
          type: string
          enum: [moving, stopped, offline]
        imei:
          type: string
        speedLimit:
          type: number
        groupName:
          type: string
        groupId:
          type: integer
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - status
        - imei
        - createdAt
        - updatedAt

    CreateDeviceBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [vehicle, asset, person, drone]
        imei:
          type: string
        speedLimit:
          type: number
        groupName:
          type: string
        groupId:
          type: integer
        userId:
          type: integer
      required:
        - name
        - type
        - imei

    UpdateDeviceBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [vehicle, asset, person, drone]
        speedLimit:
          type: number
        groupName:
          type: string
        groupId:
          type: integer
        userId:
          type: integer

    AssignDeviceBody:
      type: object
      properties:
        groupId:
          type: integer
          nullable: true
        userId:
          type: integer
          nullable: true

    Position:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        speed:
          type: number
        heading:
          type: number
        altitude:
          type: number
        accuracy:
          type: number
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - latitude
        - longitude
        - speed
        - heading
        - timestamp

    LivePosition:
      type: object
      properties:
        deviceId:
          type: integer
        deviceName:
          type: string
        deviceType:
          type: string
          enum: [vehicle, asset, person, drone]
        status:
          type: string
          enum: [moving, stopped, offline]
        latitude:
          type: number
        longitude:
          type: number
        speed:
          type: number
        heading:
          type: number
        timestamp:
          type: string
          format: date-time
        groupName:
          type: string
      required:
        - deviceId
        - deviceName
        - deviceType
        - status
        - latitude
        - longitude
        - speed
        - heading
        - timestamp

    IngestPositionBody:
      type: object
      properties:
        deviceId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        speed:
          type: number
        heading:
          type: number
        altitude:
          type: number
        accuracy:
          type: number
        timestamp:
          type: string
          format: date-time
      required:
        - deviceId
        - latitude
        - longitude
        - speed
        - heading

    Alert:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        deviceName:
          type: string
        type:
          type: string
          enum: [overspeed, offline, geofence]
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        acknowledged:
          type: boolean
        createdAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - type
        - message
        - severity
        - acknowledged
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalDevices:
          type: integer
        movingCount:
          type: integer
        stoppedCount:
          type: integer
        offlineCount:
          type: integer
        activeAlerts:
          type: integer
        vehicleCount:
          type: integer
        assetCount:
          type: integer
        personCount:
          type: integer
        droneCount:
          type: integer
        totalPositionsToday:
          type: integer
      required:
        - totalDevices
        - movingCount
        - stoppedCount
        - offlineCount
        - activeAlerts
        - vehicleCount
        - assetCount
        - personCount
        - droneCount
        - totalPositionsToday

    ActivityEvent:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        deviceName:
          type: string
        type:
          type: string
          enum: [position_update, status_change, alert_triggered, alert_acknowledged]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - deviceName
        - type
        - description
        - timestamp

    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        color:
          type: string
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt
        - updatedAt

    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
      required:
        - name
        - color

    UpdateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        groupId:
          type: integer
        groupName:
          type: string
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt

    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        groupId:
          type: integer
      required:
        - name
        - email

    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        groupId:
          type: integer